Implement fallthrough kernel.
This adds c10::fallthrough_kernel which is a special boxed function which
can be used to implement fallthrough behavior at a dispatch key.  A fallthrough
kernel will redispatch to the next valid dispatch key.  It is implemented
in such a way that it costs no more to fallthrough than it does to go
straight to the actual implementation of the kernel.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

ghstack-source-id: 8ebaf1d980b8394ce611623679ec9a9c9dbbd005
Pull Request resolved: https://github.com/pytorch/pytorch/pull/32439